+++ nix/build-bom.nix
@@ -10,16 +10,19 @@ drv:
 {
   extraPaths ? [ ],
   includeBuildtimeDependencies ? false,
+  excludes ? [ ],
 }:
 
 let
-  flags = lib.optionals includeBuildtimeDependencies [
-    "--include-buildtime-dependencies"
-  ];
+  args =
+    lib.optionals includeBuildtimeDependencies [
+      "--include-buildtime-dependencies"
+    ]
+    ++ lib.optionals (excludes != [ ]) (lib.map (e: "--exclude ${e}") excludes);
 in
 runCommand "${drv.name}.cdx.json" { nativeBuildInputs = [ transformer ]; } ''
   bombon-transformer ${drv} \
-    ${toString flags} \
+    ${toString args} \
     ${buildtimeDependencies drv extraPaths} \
     ${runtimeDependencies drv extraPaths} \
     $out